% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tirage_loto.R
\name{tirage_loto}
\alias{tirage_loto}
\title{tirage du loto}
\usage{
tirage_loto()
}
\value{
un vecteur avec 6 numéros + 2 numéros complémentaires
}
\description{
tirage du loto
}
\examples{
tirage_loto()
}
